Separate extraction of matches and non-matches
extract_proteins_cluster_neighbourhood is now changed to extract separately matches in the cluster, "gap"=non-matches in the cluster+neighbourhood, and also files to extract everything are kept. That is needed to later update the old profiles (keeping the same representatives) and make new profiles, and not mix all of this.